requirements.txt
isodate
requests
feedparser
python-dateutil
configparser